Unbreak mkdir() tracking
When the user creates a new directory, we should start tracking it
too.  The existing code forgot to add a watch for the directory
itself, however.  It also did not respect ignore rules.  Make it so.

Also move the corresponding block further up, before the event-in-dir
vs. dir-itself filter kicks in.